+++ lib/src/exports/extensions/flutter-widgets/sizedbox_ext.dart
@@ -13,12 +13,13 @@ extension SizedBoxExt on num {
 }
 
 extension SizedBoxExtWidget on Widget {
-  /// The `szdbox` function returns a `SizedBox` widget with the specified height, width, and child.
-  ///
-  /// Args:
-  ///   h (double): The height of the SizedBox.
-  ///   w (double): The parameter "w" in the above code represents the desired width of the SizedBox
-  /// widget.
-  SizedBox szdbox({double? h, double? w}) =>
-      SizedBox(height: h, width: w, child: this);
+  /// Returns:
+  ///   A custom widget named `szdbox` is being returned. This widget either returns a Container with the
+  ///   specified color, height, width, and child widget if the color is not null, or returns a SizedBox
+  ///   with the specified height, width, and child widget if the color is null.
+  Widget szdbox({double? h, double? w, Color? col}) {
+    return col != null
+        ? Container(color: col, height: h, width: w, child: this)
+        : SizedBox(height: h, width: w, child: this);
+  }
 }